COE Query Centrilization.
