+++ sources/packages/backend/apps/api/src/route-controllers/confirmation-of-enrollment/confirmation-of-enrollment.institutions.controller.ts
@@ -108,9 +108,7 @@ export class ConfirmationOfEnrollmentInstitutionsController extends BaseControll
     return {
       results: disbursementPaginatedResult.results.map(
         (disbursement: DisbursementSchedule) => {
-          const offering =
-            disbursement.studentAssessment.application.currentAssessment
-              .offering;
+          const offering = disbursement.studentAssessment.offering;
           return {
             applicationNumber:
               disbursement.studentAssessment.application.applicationNumber,
