+++ sources/packages/backend/libs/integrations/src/services/disbursement-schedule/disbursement-schedule.service.ts
@@ -1,12 +1,10 @@
 import { Injectable } from "@nestjs/common";
+import { ConfirmationOfEnrollmentService } from "@sims/services";
 import {
   RecordDataModelService,
   DisbursementSchedule,
-  ApplicationStatus,
-  COEStatus,
   OfferingIntensity,
 } from "@sims/sims-db";
-import { addDays, COE_WINDOW } from "@sims/utilities";
 import { DataSource } from "typeorm";
 
 /**
@@ -14,7 +12,10 @@ import { DataSource } from "typeorm";
  */
 @Injectable()
 export class DisbursementScheduleService extends RecordDataModelService<DisbursementSchedule> {
-  constructor(readonly dataSource: DataSource) {
+  constructor(
+    readonly dataSource: DataSource,
+    private readonly confirmationOfEnrollmentService: ConfirmationOfEnrollmentService,
+  ) {
     super(dataSource.getRepository(DisbursementSchedule));
   }
 
@@ -75,58 +76,17 @@ export class DisbursementScheduleService extends RecordDataModelService<Disburse
   async getInstitutionEligiblePendingEnrolments(): Promise<
     DisbursementSchedule[]
   > {
-    const coeThresholdDate = addDays(COE_WINDOW);
-    return this.repo
-      .createQueryBuilder("disbursementSchedule")
-      .select([
-        "disbursementSchedule.id",
-        "disbursementSchedule.disbursementDate",
-        "disbursementValues.id",
-        "disbursementValues.valueAmount",
-        "disbursementValues.valueCode",
-        "studentAssessment.id",
-        "offering.id",
-        "offering.studyStartDate",
-        "offering.studyEndDate",
-        "location.id",
-        "location.institutionCode",
-        "application.id",
-        "application.applicationNumber",
-        "application.studentNumber",
-        "student.id",
-        "student.birthDate",
-        "sinValidation.id",
-        "sinValidation.sin",
-        "user.id",
-        "user.firstName",
-        "user.lastName",
-      ])
-      .innerJoin(
-        "disbursementSchedule.disbursementValues",
-        "disbursementValues",
-      )
-      .innerJoin("disbursementSchedule.studentAssessment", "studentAssessment")
-      .innerJoin("studentAssessment.offering", "offering")
-      .innerJoin("studentAssessment.application", "application")
-      .innerJoin("offering.institutionLocation", "location")
-      .innerJoin("application.student", "student")
-      .innerJoin("student.sinValidation", "sinValidation")
-      .innerJoin("student.user", "user")
-      .where("studentAssessment.id = application.currentAssessment.id")
-      .andWhere("application.applicationStatus IN (:...status)", {
-        status: [ApplicationStatus.Enrolment, ApplicationStatus.Completed],
-      })
-      .andWhere("disbursementSchedule.hasEstimatedAwards = true")
-      .andWhere("disbursementSchedule.disbursementDate <= :coeThresholdDate", {
-        coeThresholdDate,
-      })
-      .andWhere("disbursementSchedule.coeStatus = :required", {
-        required: COEStatus.required,
-      })
+    const eligibleCOEQuery =
+      this.confirmationOfEnrollmentService.getDisbursementForCOEQuery(
+        this.repo,
+        true,
+      );
+    return eligibleCOEQuery
       .andWhere("offering.offeringIntensity = :fullTime", {
         fullTime: OfferingIntensity.fullTime,
       })
       .andWhere("location.hasIntegration = TRUE")
+      .orderBy("disbursementSchedule.id", "ASC")
       .getMany();
   }
 }
